(trunk) #3901 "Confusing error message when libtool is missing" -- fixed.
autogen.sh invokes autoreconf without checking to see if it fails (as it does if libtool can't be found). In such cases it also invokes a broken doomed version of the configure script, which generates a lot of red herring messages that distract from the real problem encountered by autoreconf.

The fix is to exit autogen.sh if its call to autoreconf fails.